refactoring propTypes and reviews page
