+++ src/components/film-page/film-page-tab-details/film-page-tab-details.test.js
@@ -4,16 +4,15 @@ const film = {
   id: `573489`,
   title: `Fantastic Beasts: The Crimes of Grindelwald`,
   poster: `fantastic-beasts-the-crimes-of-grindelwald.jpg`,
+  previewImage: `the-grand-budapest-hotel-poster.jpg`,
+  bgImage: `path`,
   releaseYear: 1999,
   genre: `Drama`,
   rating: 8.2,
   ratingVotes: 197,
   director: `Some One`,
   duration: 123,
-  description: [
-    `Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum. Sed blandit, eros vel aliquam faucibus, purus ex euismod diam, eu luctus nunc ante ut dui.`,
-    `Aliquam erat volutpat. Nunc fermentum tortor ac porta dapibus. In rutrum ac purus sit amet tempus.`,
-  ],
+  description: `Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum. Sed blandit, eros vel aliquam faucibus, purus ex euismod diam, eu luctus nunc ante ut dui. Aliquam erat volutpat. Nunc fermentum tortor ac porta dapibus. In rutrum ac purus sit amet tempus.`,
   actors: [
     `Robert Rodrigues`,
     `Takeshi Kitano`,
@@ -35,7 +34,8 @@ const film = {
       date: new Date(4231284),
       comment: `Lorem ipsum dolor sit amet, consectetur adipiscing elit. Cras aliquet varius magna`,
     }
-  ]
+  ],
+  video: `path`,
 };
 
 it(`Should FilmPageTabDetails render correctly`, () => {
