+++ src/components/films-list/films-list.jsx
@@ -1,4 +1,5 @@
 import FilmCard from "@components/film-card/film-card";
+import {Film} from "@root/types";
 
 const FilmsList = (props) => {
   const {
@@ -30,20 +31,7 @@ FilmsList.propTypes = {
   activeFilmCardId: PropTypes.string.isRequired,
   onFilmCardMouseEnter: PropTypes.func.isRequired,
   onFilmCardMouseLeave: PropTypes.func.isRequired,
-  films: PropTypes.arrayOf(PropTypes.shape({
-    id: PropTypes.string.isRequired,
-    title: PropTypes.string.isRequired,
-    poster: PropTypes.string.isRequired,
-    releaseYear: PropTypes.number.isRequired,
-    genre: PropTypes.string.isRequired,
-    rating: PropTypes.number.isRequired,
-    ratingVotes: PropTypes.number.isRequired,
-    director: PropTypes.string.isRequired,
-    description: PropTypes.string.isRequired,
-    actors: PropTypes.arrayOf(PropTypes.string.isRequired).isRequired,
-    preview: PropTypes.string.isRequired,
-    duration: PropTypes.number.isRequired,
-  }).isRequired).isRequired,
+  films: PropTypes.arrayOf(Film).isRequired,
 };
 
 export default FilmsList;
