+++ src/components/main/main.jsx
@@ -4,6 +4,7 @@ import ShowMoreButton from "@components/show-more-button/show-more-button";
 import withActiveFilmCard from "@root/hocs/with-active-film-card/with-active-film-card";
 import Footer from "@components/footer/footer";
 import Header from "@components/header/header.connect";
+import {Film} from "@root/types";
 
 const FilmsListWrapped = withActiveFilmCard(FilmsList);
 
@@ -102,29 +103,8 @@ const Main = (props) => {
 };
 
 Main.propTypes = {
-  promo: PropTypes.shape({
-    title: PropTypes.string.isRequired,
-    genre: PropTypes.string.isRequired,
-    releaseYear: PropTypes.number.isRequired,
-    bgImage: PropTypes.string.isRequired,
-    previewImage: PropTypes.string.isRequired,
-    duration: PropTypes.number.isRequired,
-    video: PropTypes.string.isRequired,
-  }).isRequired,
-  films: PropTypes.arrayOf(PropTypes.shape({
-    id: PropTypes.string.isRequired,
-    title: PropTypes.string.isRequired,
-    poster: PropTypes.string.isRequired,
-    releaseYear: PropTypes.number.isRequired,
-    genre: PropTypes.string.isRequired,
-    rating: PropTypes.number.isRequired,
-    ratingVotes: PropTypes.number.isRequired,
-    director: PropTypes.string.isRequired,
-    description: PropTypes.string.isRequired,
-    actors: PropTypes.arrayOf(PropTypes.string.isRequired).isRequired,
-    preview: PropTypes.string.isRequired,
-    duration: PropTypes.number.isRequired,
-  }).isRequired).isRequired,
+  promo: Film,
+  films: PropTypes.arrayOf(Film).isRequired,
   showingFilmsCount: PropTypes.number.isRequired,
   isPlayerActive: PropTypes.bool.isRequired,
   renderPlayer: PropTypes.func.isRequired,
